feat: Add favorite button
Add a favorite button to restaurant details to mark/unmark that restaurant
as favorite.

Also an icon is added to index to indicate that restaurant is favorite.

Aria labels are updated properly.

Data is also sent to server.

Restaurant has been sanitized as it is retrieved from server, sometimes
it is a string instead of boolean :_(