feat: Allow users to add their own reviews
Add reviews form and style it. The rating part it would be better with
a five stars widget. Form has labels for screen readers and tab order
is right. It also has some HTML5 validations. Once form is submited, it
is hidden to avoid duplication :).

Extend dbhelper (writeReview) to send proper request.